requirements.txt
# Core system dependencies
lxml~=5.4.0
PyYAML~=6.0.2
beautifulsoup4~=4.13.4
Markdown~=3.8
requests~=2.32.3
pydantic~=2.11.4

# Document parsers
python-docx~=1.1.2
openpyxl~=3.1.5
pymupdf~=1.25.5
python-pptx~=1.0.2

# Utilities
python-dateutil~=2.9.0
jsonpath-ng~=1.7.0
python-dotenv~=1.1.0
wcmatch~=10.0

# Service
flask~=3.1.0
flask_cors~=5.0.1
flask_limiter~=3.12
werkzeug~=3.1.3

# ----------------------------------------
# OPTIONAL DEPENDENCIES
# To install optional dependencies, use:
# pip install -e ".[group-name]"
# For example: pip install -e ".[db-core]"
# ----------------------------------------

# Development tools [development]
# pytest~=8.3.5
# build~=1.2.2

# Core database dependencies [db-core]
# SQLAlchemy~=2.0.40
# For macOS/Linux systems, also install:
# sqlean.py~=3.47.0; platform_system == "Darwin" or (platform_system == "Linux" and platform_machine == "x86_64")

# Scientific/Math [scientific]
# numpy~=2.0.2

# Embedding providers
# HuggingFace [huggingface]
# torch==2.7.0
# sentence-transformers~=4.1.0

# OpenAI [openai]
# openai~=1.76.0

# Anthropic [anthropic]
# anthropic~=0.34.0

# FastEmbed [fastembed]
# fastembed>=0.1.0

# Additional document parsing utilities [document_parsing]
# python-magic~=0.4.27

# Database backends
# MongoDB [db-mongodb]
# pymongo~=4.12.0

# PostgreSQL [db-postgresql]
# psycopg2-binary~=2.9.10

# MySQL [db-mysql]
# mysql-connector-python~=9.3.0

# Neo4j [db-neo4j]
# neo4j~=5.28.1

# libSQL [db-libsql]
# libsql-client~=0.3.1

# SOLR [db-solr]
# pysolr~=3.10.0

# Database vector extensions [db-vector]
# pgvector>=0.2.0
# sqlite-vec>=0.1.0
# sqlite-vss>=0.1.0

# Cloud storage [cloud-aws]
# boto3~=1.38.3
# botocore~=1.38.4

# Content Sources
# Database [source-database]
# SQLAlchemy~=2.0.40

# Confluence & JIRA [source-confluence, source-jira]
# atlassian-python-api~=3.41.9

# ServiceNow [source-servicenow]
# pysnow~=0.7.17

# MongoDB [source-mongodb]
# pymongo~=4.12.0

# ElasticSearch [source-elasticsearch]
# elasticsearch~=9.0.1

# SharePoint [source-sharepoint]
# Office365-REST-Python-Client~=2.5.0

# Google Drive [source-gdrive]
# google-api-python-client~=2.36.0
# google-auth-oauthlib~=1.0.0

# DuckDB parquet datasets [source-duckdb]  
# duckdb~=1.1.0

# ----------------------------------------
# INSTALLATION NOTE:
# For minimal installation: pip install -e .
# For full installation: pip install -e ".[all]"
# ----------------------------------------

numpy~=2.2.5
psycopg2-binary~=2.9.10
flask-cors~=5.0.1
fastembed~=0.6.1
protobuf~=6.30.2
pytest~=8.3.5

pymongo~=4.13.0
SQLAlchemy~=2.0.41
Flask-Limiter~=3.12
datefinder~=0.7.3

# Settings UI dependencies
pandas~=2.3.2
cryptography~=45.0.7
psutil~=7.0.0
